% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{add_rules}
\alias{add_rules}
\title{Install textlint rule modules}
\usage{
add_rules(rules = NULL)
}
\arguments{
\item{rules}{the name of rule; see \code{\link[=rule_sets]{rule_sets()}} and \url{https://github.com/textlint/textlint/wiki/Collection-of-textlint-rule}.}
}
\description{
Install textlint rule modules
}
\examples{
\dontrun{
add_rules("first-sentence-length")
# Skip already exist are ignored
add_rules(c("no-todo", "first-sentence-length"))
}
}
