% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_dependencies_available}
\alias{check_dependencies_available}
\title{Check textlint running environment}
\usage{
check_dependencies_available(.node_module_path = NULL)
}
\arguments{
\item{.node_module_path}{By default, it refers to the global path
and local directory. If an arbitrary path is given, it refers to that path.}
}
\description{
Check textlint running environment
}
\examples{
\dontrun{
is_available_textlint()
}
}
