% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_available_textlint}
\alias{is_available_textlint}
\title{Test for the existence of textlint as a prerequisite}
\usage{
is_available_textlint(.node_module_path = NULL)
}
\arguments{
\item{.node_module_path}{By default, it refers to the global path
and local directory. If an arbitrary path is given, it refers to that path.}
}
\description{
Test for the existence of textlint as a prerequisite
}
\examples{
\dontrun{
is_available_textlint()
}
}
