% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rule_sets}
\alias{rule_sets}
\title{Available rule names}
\usage{
rule_sets(rules = NULL)
}
\arguments{
\item{rules}{the name of rule; see \code{\link[=rule_sets]{rule_sets()}} and \url{https://github.com/textlint/textlint/wiki/Collection-of-textlint-rule}.}
}
\description{
Available rule names
}
\examples{
rule_sets(rules = c("common-misspellings", "preset-jtf-style"))
}
