% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textlint.R
\name{textlint}
\alias{textlint}
\title{Textlint a given file}
\usage{
textlint(file = NULL, lintrc = ".textlintrc", markers = TRUE)
}
\arguments{
\item{file}{filename whose target to textlint}

\item{lintrc}{file path to .textlintrc. Default, searcing from current directory.}

\item{markers}{modified output format. If \code{true}, the result of lint is
displayed in RStudio's marker panel (Only when running with RStudio version higher
than 0.99.225).}
}
\description{
Textlint a given file
}
\examples{
\dontrun{
lint_target <-
  system.file("sample.md", package = "textlintr")
textlint(lint_target)
}
}
