% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{configure_lint_rules}
\alias{configure_lint_rules}
\title{Config textlint rule file}
\usage{
configure_lint_rules(lintrc = ".textlintrc", open = FALSE)
}
\arguments{
\item{lintrc}{file path to .textlintrc. Default, searcing from current directory.}

\item{open}{If you want to modify a \code{.textlintrc}, set \code{TRUE}. (deafult \code{FALSE})}
}
\value{
list
}
\description{
Config textlint rule file
}
\examples{
\dontrun{
configure_lint_rules()
}
}
