% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textlintrc.R
\name{update_lint_rules}
\alias{update_lint_rules}
\title{Update .textlintrc}
\usage{
update_lint_rules(rules = "common-misspellings")
}
\arguments{
\item{rules}{lint rule}
}
\description{
Update .textlintrc
}
\examples{
\dontrun{
update_lint_rules(rules = "common-misspellings")
}
}
