% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textlintrc.R
\name{init_textlintr}
\alias{init_textlintr}
\title{Initialise a textlint}
\usage{
init_textlintr(rules = "common-misspellings")
}
\arguments{
\item{rules}{lint rule}
}
\description{
Initialise a textlint
}
\examples{
\dontrun{
init_textlintr()
}
}
