% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textlint.R
\name{textlint}
\alias{textlint}
\title{Textlint a given file}
\usage{
textlint(file = NULL, lintrc = ".textlintrc")
}
\arguments{
\item{file}{filename whose target to textlint}

\item{lintrc}{file path to .textlintrc. Default, searcing from current directory.}
}
\description{
Textlint a given file
}
\examples{
\dontrun{
lint_target <-
  system.file("sample.md", package = "textlintr")
textlint(lint_target)
}
}
