% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textlintrc.R
\name{update_lint_rules}
\alias{update_lint_rules}
\title{Update .textlintrc}
\usage{
update_lint_rules(rules = NULL, lintrc = ".textlintrc",
  overwrite = FALSE)
}
\arguments{
\item{rules}{the name of rule; see \code{\link[=rule_sets]{rule_sets()}} and \url{https://github.com/textlint/textlint/wiki/Collection-of-textlint-rule}.}

\item{lintrc}{file path to .textlintrc. Default, searcing from current directory.}

\item{overwrite}{logical. If set \code{TRUE}, existing rules will be overwritten by the input rules.}
}
\description{
Update the rule file (\code{.texlintrc}) which textlint checks.
To adopt the rule of the character string specified as argument. When
\code{NULL} is given, all installed rules are applied.
}
\examples{
\dontrun{
# Registrate all installed rules
update_lint_rules()
# Added a rule
update_lint_rules(rules = "common-misspellings", overwrite = FALSE)
# Overwite rules
update_lint_rules(rules = "common-misspellings", overwrite = TRUE)
}
}
